+++ src/app/mod.rs
@@ -2,5 +2,6 @@ pub mod about;
 pub mod cli;
 pub mod client_settings;
 pub mod logger;
+pub mod mpv_setup;
 pub mod settings;
 pub mod updater;